refactor: 💡 dynamic import for framer motion
