+++ src/components/DataRoom/IndustryComparison/index.tsx
@@ -1,12 +1,11 @@
 import type { BaseBlock } from '@/components/Home/types'
-import { useTransform, motion, useScroll } from 'framer-motion'
+import { useScroll } from 'framer-motion'
 import dynamic from 'next/dynamic'
 import React, { useRef } from 'react'
 import css from './styles.module.css'
 import { useIsMediumScreen } from '@/hooks/useMaxWidth'
-import { Typography } from '@mui/material'
-import DotGrid from './DotGrid'
 
+const Content = dynamic(() => import('./Content'))
 const SlidingPanel = dynamic(() => import('@/components/common/SlidingPanel'))
 
 const IndustryComparison = ({ title }: BaseBlock) => {
@@ -19,9 +18,6 @@ const IndustryComparison = ({ title }: BaseBlock) => {
     offset: ['start end', 'end start'],
   })
 
-  const opacityParams = [0.25, 0.35, 0.65, 0.75]
-  const opacity = useTransform(scrollYProgress, opacityParams, [0, 1, 1, 0])
-
   return (
     <div ref={backgroundRef} className={css.sectionContainer}>
       <div ref={gridContainerRef} className={css.stickyContainer}>
@@ -30,17 +26,7 @@ const IndustryComparison = ({ title }: BaseBlock) => {
           translateParams={isMobile ? ['0%', '0%'] : ['100%', '0%', '0%', '100%']}
           scrollYProgress={scrollYProgress}
         >
-          <motion.div
-            className={css.slidingPanelContent}
-            style={{
-              opacity: isMobile ? 1 : opacity,
-            }}
-          >
-            <Typography className={css.title} variant="h1">
-              {title}
-            </Typography>
-            <DotGrid containerRef={gridContainerRef} scrollYProgress={scrollYProgress} />
-          </motion.div>
+          <Content title={title} containerRef={gridContainerRef} scrollYProgress={scrollYProgress} />
         </SlidingPanel>
       </div>
     </div>
